Fixed some compiler warnings on W32

+++ src/microhttpd/mhd_sockets.h
@@ -443,8 +443,8 @@ typedef int MHD_SCKT_SEND_SIZE_;
       (((void*) (w) == (void*) 0) || ((fd_set*) (w))->fd_count == 0) &&  \
       (((void*) (e) == (void*) 0) || ((fd_set*) (e))->fd_count == 0) ) ? \
     ( ((void*) (t) == (void*) 0) ? 0 :                                   \
-      (Sleep (((struct timeval*) (t))->tv_sec * 1000                     \
-              + ((struct timeval*) (t))->tv_usec / 1000), 0) ) :         \
+      (Sleep ((DWORD)((struct timeval*) (t))->tv_sec * 1000              \
+              + (DWORD)((struct timeval*) (t))->tv_usec / 1000), 0) ) :  \
     (select ((int) 0,(r),(w),(e),(t))) )
 #endif
